requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
python-decouple==3.8
python-ldap==3.4.3
drf-api-tracking==1.8.0
weasyprint==58.1
qrcode==7.4.2
requests==2.28.2
Unidecode==1.3.6
drf-flex-fields==1.0.2
zeep==4.2.1
validate-docbr==1.10.0